# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Photo Store API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: albums
    description: Album operations
  - name: photos
    description: Photo operations
  - name: orders
    description: Order and checkout operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /albums:
    get:
      operationId: getAlbums
      tags: [albums]
      summary: List published albums
      responses:
        "200":
          description: List of published albums
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Album"

  /albums/{slug}:
    get:
      operationId: getAlbumBySlug
      tags: [albums]
      summary: Get a published album with its photos
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Album with photos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumWithPhotos"
        "404":
          description: Album not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /checkout:
    post:
      operationId: createCheckout
      tags: [orders]
      summary: Create a Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Stripe checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Album not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/albums:
    get:
      operationId: adminGetAlbums
      tags: [admin]
      summary: List all albums (admin)
      security:
        - AdminSecret: []
      responses:
        "200":
          description: All albums
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminAlbum"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: adminCreateAlbum
      tags: [admin]
      summary: Create a new album
      security:
        - AdminSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlbumRequest"
      responses:
        "201":
          description: Created album
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAlbum"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/albums/{id}:
    get:
      operationId: adminGetAlbum
      tags: [admin]
      summary: Get album with photos (admin)
      security:
        - AdminSecret: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Album with photos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumWithPhotos"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: adminUpdateAlbum
      tags: [admin]
      summary: Update an album
      security:
        - AdminSecret: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlbumRequest"
      responses:
        "200":
          description: Updated album
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAlbum"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteAlbum
      tags: [admin]
      summary: Delete an album
      security:
        - AdminSecret: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/albums/{id}/publish:
    patch:
      operationId: adminPublishAlbum
      tags: [admin]
      summary: Publish or unpublish an album
      security:
        - AdminSecret: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishAlbumRequest"
      responses:
        "200":
          description: Updated album
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAlbum"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/albums/{id}/cover:
    patch:
      operationId: adminSetCoverPhoto
      tags: [admin]
      summary: Set the cover photo for an album
      security:
        - AdminSecret: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetCoverRequest"
      responses:
        "200":
          description: Updated album
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAlbum"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/photos/{id}:
    delete:
      operationId: adminDeletePhoto
      tags: [admin]
      summary: Delete a photo
      security:
        - AdminSecret: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/upload:
    post:
      operationId: adminUploadPhotos
      tags: [admin]
      summary: Upload photos to an album (multipart form)
      security:
        - AdminSecret: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                albumId:
                  type: string
                files:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
                - albumId
                - files
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/orders:
    get:
      operationId: adminGetOrders
      tags: [admin]
      summary: List all orders (admin)
      security:
        - AdminSecret: []
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    AdminSecret:
      type: apiKey
      in: header
      name: x-admin-secret

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    Photo:
      type: object
      properties:
        id:
          type: string
        albumId:
          type: string
        filename:
          type: string
        watermarkedUrl:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - albumId
        - filename
        - watermarkedUrl
        - sortOrder
        - createdAt

    Album:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        priceSingle:
          type: number
        priceAlbum:
          type: number
        isPublished:
          type: boolean
        photoCount:
          type: integer
        coverPhotoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - slug
        - description
        - priceSingle
        - priceAlbum
        - isPublished
        - photoCount
        - coverPhotoUrl
        - createdAt

    AdminAlbum:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        priceSingle:
          type: number
        priceAlbum:
          type: number
        isPublished:
          type: boolean
        photoCount:
          type: integer
        coverPhotoId:
          type: ["string", "null"]
        coverPhotoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - slug
        - description
        - priceSingle
        - priceAlbum
        - isPublished
        - photoCount
        - coverPhotoId
        - coverPhotoUrl
        - createdAt

    AlbumWithPhotos:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        priceSingle:
          type: number
        priceAlbum:
          type: number
        isPublished:
          type: boolean
        coverPhotoId:
          type: ["string", "null"]
        coverPhotoUrl:
          type: ["string", "null"]
        photos:
          type: array
          items:
            $ref: "#/components/schemas/Photo"
        createdAt:
          type: string
      required:
        - id
        - title
        - slug
        - description
        - priceSingle
        - priceAlbum
        - isPublished
        - coverPhotoId
        - coverPhotoUrl
        - photos
        - createdAt

    CreateAlbumRequest:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        priceSingle:
          type: number
        priceAlbum:
          type: number
        isPublished:
          type: boolean
      required:
        - title
        - slug
        - priceSingle
        - priceAlbum

    UpdateAlbumRequest:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        priceSingle:
          type: number
        priceAlbum:
          type: number

    PublishAlbumRequest:
      type: object
      properties:
        isPublished:
          type: boolean
      required:
        - isPublished

    SetCoverRequest:
      type: object
      properties:
        photoId:
          type: string
      required:
        - photoId

    CheckoutRequest:
      type: object
      properties:
        albumId:
          type: string
        purchaseType:
          type: string
          enum: [single, album]
        photoId:
          type: ["string", "null"]
      required:
        - albumId
        - purchaseType

    CheckoutResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url

    UploadResult:
      type: object
      properties:
        uploaded:
          type: array
          items:
            type: string
        errors:
          type: array
          items:
            type: string
      required:
        - uploaded
        - errors

    Order:
      type: object
      properties:
        id:
          type: string
        customerEmail:
          type: string
        albumId:
          type: string
        albumTitle:
          type: string
        purchaseType:
          type: string
        photoId:
          type: ["string", "null"]
        amountPaid:
          type: number
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - customerEmail
        - albumId
        - albumTitle
        - purchaseType
        - photoId
        - amountPaid
        - status
        - createdAt