# lib/api-zod/src/generated/types/album.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photo Store API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Album {
  id: string;
  title: string;
  slug: string;
  /** @nullable */
  description: string | null;
  priceSingle: number;
  priceAlbum: number;
  isPublished: boolean;
  photoCount: number;
  /** @nullable */
  coverPhotoUrl: string | null;
  createdAt: string;
}