# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photo Store API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: string;
  customerEmail: string;
  albumId: string;
  albumTitle: string;
  purchaseType: string;
  /** @nullable */
  photoId: string | null;
  amountPaid: number;
  status: string;
  createdAt: string;
}